Added mod_webkit for Apache 2 64bit on Windows.

+++ WebKit/Adapters/mod_webkit2/mod_webkit.c
@@ -570,7 +570,7 @@ static int webkit_handler(request_rec *r)
     /* end dictionary */
     w_byte(TYPE_NULL, whole_dict);
 
-    length = whole_dict->ptr - whole_dict->str;
+    length = (long)(whole_dict->ptr - whole_dict->str);
 
     write_integer((int)length, int_dict);
